Enables Hiero to keep track of the last used paths, glyphs and filenames
This is done on a per-action basis so that each type of directory is
stored/restored separately: the actual Preferences instance is being
used to store this information persistently.

Also keeps track of the last filename used to export .hiero/.fnt files: the
first time the save dialog is opened, a filename is suggested.