Fix touch events for LWP on Android API <9
Without this fix HCT Hero with 2.2.1 does not respond touch events.
In removed "FIXME" comment stated about crash on ICS - i check it on
the ICS 4.0.3 and all works, but anyway - i add SDK check.
On the LG P990 with 2.3.5 no problems without SDK check.
And works on the Samsung S5670 with 2.3.6. So, hope it's usefull.