Add 'includeantruntime="false"' to javac
Newer versions of ANT warn about the fact that the ANT runtime is
included on a javac compile.  Since this is not generally what is
wanted (but is backwards compatible with older versions of ANT), a new
flag
	includeantruntime="false"
will disable it.

This will clear up the Jenkins warnings like this:

    [javac] /var/lib/jenkins/workspace/libgdx/build-template.xml:51: warning: 'includeantruntime' was not set, defaulting to build.sysclasspath=last; set to false for repeatable builds

AFAICT, this is backwards compatible with older versions of ANT ...