Refactored parentToLocalCoordinates.
I tested this via GroupTest and it seems to work. Not sure why we were doing two rotations for origin? If there was a good reason I'd like to include a test case in GroupTest. I wrote localToParentCoordinates with only a single rotation, so if we really do need two rotations that would also be broken.

Besides using a single rotation, I removed branching where we only saved a couple subtractions.