Add classpath files support for gwt (limited)
Classpath files, such as the default bitmap font and the default shaders
for the new 3d api, are now available in gwt. This limited support
requires the classpath files to be specified in the gwt module definition
by setting the multi-valued property 'gdx.files.classpath'. Any file
specified will be copied to the assets folder at compile time.