SelectBox: added set max # items to display in dropdown
Added backwards compatible functionality for configuring the max # of items to display in the dropdown SelectList.

You set the maximum number of items you want displayed in the dropdown box until scrolling is required, or <= 0 to have it list them all, but should it reach the bottom/top of the screen then scrolling is required. 

This is also a solution to an API defect. There was no previous easy way to modify the inner SelectList's prefHeight() to adjust the dropdown list's displayed height.