Fix bug in IntFloatMap, IntMap, LongMap get
The zeroValue was incorrectly being returned instead of the default value
when hasZeroValue was false. This had the effect of returning the last set
zeroValue even after it has been unset. Not so much an issue for IntMap and
LongMap because the zeroValue was nullified on remove, but added check for
consistency and clarity.