Update custom mouse cursor image support.
The Windows operating system has the additional requirement for a custom
mouse cursor image to have a width and height that are both
powers-of-two greater than zero (not necessarily equal).

Update the documentation of the setCursorImage method in the Gdx.Input
interface to add this requirement.

Update the setCursorImage method to throw a GdxRuntimeException if the
width or height are not both powers-of-two greater than zero.

Improve exception messages that occur when a hotspot param passed to
Gdx.Input.setCursorImage is outside the bounds of the image represented
by the pixmap param.