Added support for linking with native libs on iOS.
This is done by pulling down a jar ("natives-ios" as a compile
dependency), and marking it both as a compile and natives dependecy in
the project ":ios" section in the main build.gradle template.

The actual extraction and copying is done by the copyNatives task in
ios/build.gradle. After all the .a files have been extracted, the
buildRoboVMXML task in the same build.gradle file creates the necessary
lib entries in the libs section of a fresh robovm.xml file. The same
task also adds the assets directory to robovm.xml (only one asset
directory is currently supported, but everything in it and beneath gets
pulled into the iOS package).

I have tested the gradle robovm plugins' use of robovm.properties as
well. By simply creating a robovm.properties file in ios/ the tasks from
the gradle robovm plugin will detect it and pull it in. I've tried
setting the app.name in it to control the name of the app (as shown
below the launch icon), but since I was unsuccesful in changing the name
I haven't even tried adding additional support for the name below the
icon yet.