Removed null checks for Cell values.
The values are only null to indicate the value is not set. This only occurs for column, row and cell defaults. Once the cell values are merged with the defaults, none of the values should be null. The getters are only useful to call after the cell has been merged with the defaults.