Update to mapping file for the Ouya
Updated mapping as some values were incorrect. (L1 and L2 were swapped,
same with R1 and R2). Changes tested on a real device.

Removed conditional mapping, as I don't think it is a good idea.
Ouya.BUTTON_O should always return the same value in my opinion,
previously it would return one value if it running on an Ouya device,
another if it wasn't.

Adding the ability to "simulate" running on an Ouya device for testing
purposes. E.g. In my application I show device specific artwork and I
mainly test on my PC. I was careful to maintain backwards compatibility
with this, so it will not break any one's build who was using the
"runningOnOuya" boolean