Refactored scene2d stage debug rendering.
It is no longer required to manage a debug renderer and call draw, stage has one internally. Like the old table debug rendering, there is no overhead if debug is not turned on for any actors. The debug renderer can still be used on its own, but I'm not sure anyone will ever need to do that. The rest of the changes were mostly just shuffling things around to match the rest of the scene2d organization.

I wonder if it would be more powerful for actors to have a `drawDebug(ShapeRenderer)` method? Actors could draw more than rects (lines, circles, Xs, etc). Table is the only one that caches its rects and it could do that internally. `Group#computeTransform` can be used to adjust the transform matrix for the ShapeRenderer so `drawDebug` can use it with local coordinates, same as how it is done for `SpriteBatch`. The downside is that this causes a ShapeRenderer batch flush per group, but then this is already for debugging so performance shouldn't be an issue. What do you guys think?

This was a nice addition to scene2d, kudos to @nooone!

#2011