Make iPhone 6+ behave more sensibly.
The iPhone 6+ behaves badly without this.  It returns a density of 1 which makes apps think its a very very large screen (vs a regular screen with very small pixels). Note that this also affects ppiX & ppiY (et al), but in an appropriate way.

This is not a perfect fix for density on iOS, but it's a decent bandaid for now.