Added FilehandleResolver for classpaths
Case: The resolver will be useful for libraries that use Libgdx as a backend and have default assets packed in the library in case a desired asset can't be found to draw/play/etc.

The InternalFileHandleResolver does not suffice, as it prioritizes looking in the internal storage first. From the documentation: "If a file can't be found on the internal storage, the file module falls back to searching the file on the classpath". In my case ClasspathFileHandleResolver will be used because we do not want to potentially use assets with the same name that the user has placed in the assets folder.

Example: Have a "Missing texture" texture drawn when developer using my library want to draw a texture from a filehandle that does not exist. The "Missing texture" could in this case be located at MissingTexture.png in the classpath. If the developer using the library already has a MissingTexture.png file in internal storage, that one would be used instead if we resolve the file using InternalFileHandleResolver. We don't want the developer using my library have to know by heart file names that have to be avoided.