Fix false-positive of vibration support
We get our vibrator with
context.getSystemService(Context.VIBRATOR_SERVIBE).

The previous check tested if(vibrator != null), but according to 
http://developer.android.com/reference/android/content/Context.html#getSystemService(java.lang.String)  
it returns "The service or null if the name does not exist.". So it
should only return null if the name does not exist, which should be
never the case.
Instead of that we now use vibrator.hasVibrator